Update the prior probability of patterns
This ensures that sum of prior probabilities over all (entirely and partially observed) patterns equal 1 while preserving the fact than patterns of length i are 1/3 less probable than patterns of length i+1.
For trees of depth larger or equal to ~8, this induces differences that are so small than it does not impact the inference results.